+++ Visor.h
@@ -7,8 +7,8 @@
 class Visor
 {
 private:
-    int m_position;
-    const int m_maxAngleError;
+    double m_position;
+    const double m_maxAngleError;
     const DefinedPoint m_point;
     Vector m_vectorCoordinate;
     Angle m_angle;
@@ -16,12 +16,12 @@ private:
     Ray m_negativeErrorRay;
 public:
     // Если что добавить что ошибки разные в конструктор
-    Visor(const int trueCoordinateX, const int trueCoordinateY, const int currentPosition, const int maxAngleError);
+    Visor(const double trueCoordinateX, const double trueCoordinateY, const double currentPosition, const double maxAngleError);
     Ray GetPositiveErrorRay() const;
     Ray GetNegativeErrorRay() const;
-    int GetMaxAngleErorr() const;
-    int GetAngle() const;
-    int GetPosition() const;
+    double GetMaxAngleErorr() const;
+    double GetAngle() const;
+    double GetPosition() const;
     void Recalculate(const int position);
 };
 #endif // VISOR_H
